# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Padel Concierge API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/users:
    get:
      operationId: listAdminUsers
      tags: [admin]
      summary: List all users (admin) with optional type filter
      parameters:
        - name: userType
          in: query
          required: false
          schema:
            type: string
            enum: [real_user, seed_test, beta_tester]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
  /admin/users/{id}/user-type:
    patch:
      operationId: patchAdminUserType
      tags: [admin]
      summary: Set user type (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userType:
                  type: string
                  enum: [real_user, seed_test, beta_tester]
              required:
                - userType
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Bad request
        "403":
          description: Forbidden

  # COACH
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        level:
          type: string
        goal:
          type: string
        intensity:
          type: string
        locationLat:
          type: ["number", "null"]
        locationLng:
          type: ["number", "null"]
        locationName:
          type: ["string", "null"]
        avatar:
          type: ["string", "null"]
        verified:
          type: boolean
        verificationDate:
          type: ["string", "null"]
        role:
          type: string
        favouritePlayers:
          type: array
          items:
            type: integer
        availability:
          type: array
          items:
            $ref: "#/components/schemas/AvailabilityEntry"
        matchesPlayed:
          type: integer
        wins:
          type: integer
        language:
          type: string
        isOnline:
          type: boolean
        lastActive:
          type: ["string", "null"]
        createdAt:
          type: string
        lastLogin:
          type: ["string", "null"]
        archetype:
          type: ["string", "null"]
        userType:
          type: string
          enum: [real_user, seed_test, beta_tester]
        isBeta:
          type: boolean
        modePlayer:
          type: boolean
        modeCoach:
          type: boolean
        modeAdmin:
          type: boolean
        modeDeveloper:
          type: boolean
      required:
        - id
        - name
        - email
        - phone
        - level
        - goal
        - intensity
        - verified
        - role
        - favouritePlayers
        - availability
        - matchesPlayed
        - wins
        - language
        - isOnline
        - createdAt
        - userType
        - modePlayer
        - modeCoach
        - modeAdmin
        - modeDeveloper
    AvailabilityEntry:
      type: object
      properties:
        date:
          type: string
        slots:
          type: array
          items:
            type: array
            items:
              type: string
      required:
        - date
        - slots
...